@@ -2,26 +2,20 @@ package com.onehilltech.gatekeeper.android.examples.standard;
 
 import android.content.Intent;
 import android.os.Bundle;
+import android.os.Looper;
 import android.support.annotation.Nullable;
 import android.support.v7.app.AppCompatActivity;
-import android.view.View;
 import android.widget.Button;
 
 import com.onehilltech.backbone.http.HttpError;
 import com.onehilltech.gatekeeper.android.GatekeeperSessionClient;
 import com.onehilltech.gatekeeper.android.GatekeeperSignInActivity;
 
-import retrofit2.Call;
-import retrofit2.Callback;
-import retrofit2.Response;
-
 public class MainActivity extends AppCompatActivity
   implements GatekeeperSessionClient.Listener
 {
   private Button btnSignOut_;
   private GatekeeperSessionClient session_;
-
-  private static final String TAG = "MainActivity";
 
   @Override
   protected void onCreate (@Nullable Bundle savedInstanceState)
@@ -49,32 +43,10 @@ public class MainActivity extends AppCompatActivity
   {
     this.btnSignOut_ = (Button)this.findViewById (R.id.btn_signout);
     this.btnSignOut_.setEnabled (this.session_.isSignedIn ());
-    this.btnSignOut_.setOnClickListener (new View.OnClickListener ()
-    {
-      @Override
-      public void onClick (View v)
-      {
-        onLogoutClicked ();
-      }
-    });
-  }
 
-  private void onLogoutClicked ()
-  {
-    this.session_.signOut (new Callback<Boolean> ()
-    {
-      @Override
-      public void onResponse (Call<Boolean> call, Response<Boolean> response)
-      {
-        if (response.isSuccessful () && response.body ())
-          session_.ensureSignedIn (MainActivity.this, GatekeeperSignInActivity.class);
-      }
-
-      @Override
-      public void onFailure (Call<Boolean> call, Throwable t)
-      {
-
-      }
+    this.btnSignOut_.setOnClickListener (v -> {
+      this.session_.signOut ()
+                   .then ((value, cont) -> this.session_.ensureSignedIn (this, GatekeeperSignInActivity.class));
     });
   }
 
@@ -93,6 +65,9 @@ public class MainActivity extends AppCompatActivity
   @Override
   public void onSignedIn (GatekeeperSessionClient client)
   {
+    if (!Looper.getMainLooper ().getThread ().equals (Thread.currentThread ()))
+      System.err.println ("Thread is not the main looper");
+
     this.btnSignOut_.setEnabled (true);
   }
 
